+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.39.6] - 2020-06-15
+
 - Made default admin credentials configurable
 - Fix: Run migrations in transactions so nothing is destroyed on a faulty migration